+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * added multiple filters for the same field connected with OR
 * added support for "contains" for filter fields
+* added reading filter options using the `Analyzer`
 * refactored code into modules
 * refactored `Analyzer` to group methods
 * updated User-Agent blacklist